+++ src/common/utils/deployTips.ts
@@ -3,7 +3,7 @@ import * as c from "chalk";
 import * as format from "string-template";
 import stripAnsi = require("strip-ansi");
 import { GitProvider } from "../gitProvider";
-import { PullRequestMessageRequest } from "../gitProvider/gitProvider";
+import { PullRequestMessageRequest } from "../gitProvider/types/gitProvider";
 import { GitProviderRoot } from "../gitProvider/gitProviderRoot";
 import { getAllTips } from "./deployTipsList";
